+++ tests/models/job/test_orders.py
@@ -348,6 +348,10 @@ class TestOrderModel(BaseModelTest):
             roles[role_name],
             origin_state
         )
+        # TODO: improve this with requirement items, for now disable it
+        order.requirement_items = []
+        order.project.project_type = 'on-demand'
+        session.flush()
 
         new_requirements = {}
         with pytest.raises(WorkflowTransitionException) as excinfo:
@@ -362,6 +366,7 @@ class TestOrderModel(BaseModelTest):
         assert 'Field requirements is required for this transition' in str(excinfo)
 
         new_requirements['requirements'] = 'New requirements for this order!'
+        new_requirements['number_required_assets'] = 25
         wf.edit_requirements(fields=new_requirements)
         session.flush()