+++ tests/views/test_orders_view.py
@@ -35,10 +35,10 @@ class TestOrderView(BaseVersionedTestView):
     ]
     serialize_attrs = [
         'path', '_roles', '_actors', 'customer', 'project', 'timezone',
-        'assignment', 'assignments'
+        'assignment', 'assignments', 'requirement_items'
     ]
     ignore_validation_fields = [
-        'state_history', 'state', 'location'
+        'state_history', 'state', 'location', 'requirements', 'number_required_assets'
     ]
     file_path = 'data/orders.json'
     model = models.Order